RateLimit for Lexoffice API, fixes toarray

+++ src/Entities/Articles/Price.php
@@ -11,7 +11,7 @@ class Price extends NamedEntity {
     protected float $netPrice;
     protected float $grossPrice;
     protected LeadingPrice $leadingPrice;
-    protected int $taxRate;
+    protected float $taxRate;
 
     public function __construct($data = null) {
         parent::__construct($data);
